old-tests/submission/W3C/domhtml/HTMLObjectElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>object</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/association-of-controls-and-forms.html#dom-fae-form" rel="help">
<link rel="author" title="W3C">
</head>
<body>

<object form="form1" data="/support/html/sample.html" height="80" type="text/html" width="550"></object>
<object form="form2" data="/support/html/sample.html" height="80" type="text/html" width="550"></object>

<form action="./files/getData.pl" method="post" id="form1">
</form>



<script>

test(function() {
  var testNode = window.document.getElementsByTagName("object").item(0);
     assert_equals(testNode.form.tagName, "FORM");
     assert_equals(testNode.form.id, "form1");
}, "object.form has a form element");

test(function() {
  var testNode = window.document.getElementsByTagName("object").item(1);
  assert_equals(testNode.form, null);
}, "object.form is null");

</script>
<p>If the element has a form content attribute, and if the first element in the Document to have an ID that is case-sensitively equal to the element's form content attribute's value is a form element, then associate the form-associated element with that form element.</p>
<div id="log"></div>
</body></html>
